msm_audio_calibration: Undefine TOPOLOGY_SPECIFIC_CHANNEL_INFO on legacy
This define was erroneously inverted to _include_ all legacy socs, when
they should instead be excluded from the unsupported topology feature.
It also causes SM8150 to be excluded, which supports this feature. On
the bright side, this triggers a build error where the member fields
are omitted from a structure while code in q6voice is setting these
fields unconditionally. Said code has been omitted when
TOPOLOGY_SPECIFIC_CHANNEL_INFO is not set in a future commit.

Signed-off-by: MarijnS95 <marijns95@gmail.com>